crewai-outbound/outbound_commercial/requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
langchain_community==0.0.29
python-dotenv==1.0.1
openai>=1.40.0
tenacity==8.5.0
loguru==0.7.2
